front end things and fixes

+++ src/java/facade/CompanyInterface.java
@@ -15,7 +15,8 @@ public interface CompanyInterface
     public CityInfo getCityInformation(int zipCode);
     public Company createCompany(Company comp);
     public Company getCompany (int cvr)throws CompanyNotFoundException;
+//    public Company deleteCompanyById (long id)throws CompanyNotFoundException;
     public Company updateCompany(Company comp)throws CompanyNotFoundException;
-    public Company deleteCompany(int cvr)throws CompanyNotFoundException;
+    public Company deleteCompany(long cvr)throws CompanyNotFoundException;
     public List<Company> getCompanies();
 }
